+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.3
+
+* Provide metadata schema in JSON for the IMGAG dropbox
+* Register checksums for Oxford Nanopore datasets
+* Register unclassified read data for Oxford Nanopore datasets
+
 ## 1.2
 
 * Provide ETL routine for Oxford Nanopore NGS data